move chmod, fix typo, add some fcntl branches
